requirements.txt
# Project_AIVideoFlow - Python依赖包

# 网页数据采集
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
pillow==10.3.0

# 图像处理
opencv-python==4.9.0.80
numpy==1.26.4
scikit-image==0.23.2

# OCR相关
paddleocr==2.7.0
paddlepaddle==2.6.2
easyocr==1.7.1

# AI/ML相关
transformers==4.38.2
torch==2.2.1
torchvision==0.17.1

# API调用
openai==1.12.0
anthropic==0.25.4

# 工具库
tqdm==4.66.4
python-dotenv==1.0.1
pyyaml==6.0.1
json5==0.9.14

# 开发工具
pytest==8.1.1
black==24.2.0
flake8==7.0.0

# ComfyUI相关
comfyui-client==0.1.0
websocket-client==1.7.0